drfecommerce/requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
coverage==7.3.2
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
importlib-resources==6.1.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
pluggy==1.3.0
Pygments==2.16.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
zipp==3.17.0
